backend/pyproject.toml
[project]
name = "cartolapy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.5",
    "aiolimiter>=1.2.1",
    "fastapi[standard]>=0.135.3",
    "pandas>=3.0.2",
    "pyarrow>=15.0.0",
    "redis>=5.0.0",
    "pydantic-settings>=2.0.0",
    "slowapi>=0.1.9",
    "taskiq-fastapi>=0.4.0",
    "taskiq-redis>=0.6.0",
]

[dependency-groups]
dev = [
    "marimo>=0.22.0",
    "ruff>=0.15.8",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-env",
    "anyio>=4.0.0",
]

[tool.pytest.ini_options]
env = [
    "ENVIRONMENT=pytest",
    "REDIS_URL=redis://localhost:6379",
]
